Add bread crumbs so VMs use eth1 as part of OVS bridge
With this change, I'm leaving a hint on how I can change the vms
to use tap1 as a bridged interface. That allows me to connect
the nodes to other ports in the bridge, such as ports to docker
containers.

That is used as an example in this page:
http://www.flaviof.com/blog/work/how-to-odl-in-docker.html

Note that the hypervisor is perfectly okay with having multiple
VMs on top of the same tap1 interface, so there is no need to
create tap2, tap3... etc.